Move Network Commissioning driver source includes into app projects
Network commissioning source is unconditionally included in the Linux
platform.  But this only makes sense for commissionable apps that
actually use this code.

This commit moves these source includes to the apps and also fixes a
build error that is arising from blindly assigning size_t to a uint8_t.

+++ src/include/platform/NetworkCommissioning.h
@@ -79,7 +79,7 @@ protected:
 struct Network
 {
     uint8_t networkID[kMaxNetworkIDLen];
-    uint8_t networkIDLen;
+    size_t networkIDLen;
     bool connected;
 };
