+++ src/platform/Linux/NetworkCommissioningDriver.h
@@ -180,7 +180,7 @@ public:
 
         // Public, but cannot be accessed via NetworkIterator interface.
         uint8_t interfaceName[kMaxNetworkIDLen];
-        uint8_t interfaceNameLen = 0;
+        size_t interfaceNameLen = 0;
         bool exhausted           = false;
     };
